[FEATURE] #72 : Custom Plugin 구현 (Kotlin)
