--- client/app/src/main/java/com/example/healthc/data/source/recipe/RecipeDataSource.kt
@@ -1,11 +1,0 @@
-package com.example.healthc.data.source.recipe
-
-import com.example.healthc.data.model.remote.recipe.IngredientResponse
-import com.example.healthc.data.model.remote.recipe.NutritionLabelResponse
-import com.example.healthc.data.model.remote.recipe.RecipeResponse
-
-interface RecipeDataSource {
-    suspend fun getRecipeList(ingredient: String): Result<List<RecipeResponse>>
-
-    suspend fun getIngredientList(detectedObject : String) : Result<List<IngredientResponse>>
-}
