+++ client/build-logic/convention/src/main/java/com/healthc/plugin/AndroidKotlinPlugin.kt
@@ -1,0 +1,40 @@
+package com.healthc.plugin
+
+import com.android.build.gradle.BaseExtension
+import org.gradle.api.JavaVersion
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.gradle.api.artifacts.VersionCatalogsExtension
+import org.gradle.api.plugins.ExtensionAware
+import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.getByType
+import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
+
+class AndroidKotlinPlugin: Plugin<Project> {
+    override fun apply(target: Project) {
+        with(target){
+            with(pluginManager) {
+                apply("org.jetbrains.kotlin.android")
+            }
+
+            extensions.getByType<BaseExtension>().apply {
+                compileOptions {
+                    sourceCompatibility = JavaVersion.VERSION_17
+                    targetCompatibility = JavaVersion.VERSION_17
+                }
+
+                (this as ExtensionAware).configure<KotlinJvmOptions> {
+                    jvmTarget = "17"
+                }
+            }
+
+            val libs = extensions.getByType<VersionCatalogsExtension>().named("libs")
+
+            dependencies {
+                "implementation"(libs.findLibrary("kotlin").get())
+                "implementation"(libs.findLibrary("kotlin-coroutines").get())
+            }
+        }
+    }
+}
