+++ client/data/src/main/java/com/healthc/data/source/recipe/RecipeDataSource.kt
@@ -1,0 +1,10 @@
+package com.healthc.data.source.recipe
+
+import com.healthc.data.model.remote.recipe.IngredientResponse
+import com.healthc.data.model.remote.recipe.RecipeResponse
+
+interface RecipeDataSource {
+    suspend fun getRecipeList(ingredient: String): Result<List<RecipeResponse>>
+
+    suspend fun getIngredientList(detectedObject : String) : Result<List<IngredientResponse>>
+}
